add list view services and files
