+++ services/chile_atiende.py
@@ -1,88 +1,8 @@
-import requests
-
-from django.conf import settings
-
 from .models import Service as ServiceModel
 from .models import File as FileModel
 
-
-class ChileAtiende(object):
-
-    def __init__(self):
-        self._url = 'http://www.chileatiende.cl/api'
-        self._access_token = settings.CHILEATIENDE_ACCESS_TOKEN
-
-    def _connect(self, url):
-        headers = {
-            'User-Agent': 'Mozilla/5.0',
-        }
-        request = requests.get(
-            url,
-            headers=headers
-        )
-
-        return request
-
-    def _get_url(self, path, id=None, is_file=False):
-
-        url = '{}{}'.format(
-            self._url,
-            path,
-        )
-        if id:
-            url = '{}/{}/'.format(
-                url,
-                id,
-            )
-
-        if is_file:
-            url = '{}{}'.format(
-                url,
-                'fichas/'
-            )
-
-        url = '{}?access_token={}&type=json'.format(
-                url,
-                self._access_token,
-            )
-
-        return url
-
-
-class Service(ChileAtiende):
-
-    def list(self):
-
-        return self._connect(
-            self._get_url('/servicios')
-        )
-
-    def get(self, service_id):
-
-        return self._connect(
-            self._get_url('/servicios', service_id)
-        )
-
-
-class File(ChileAtiende):
-
-    def list(self):
-
-        return self._connect(
-            self._get_url('/fichas')
-        )
-
-    def get(self, file_id):
-
-        return self._connect(
-            self._get_url('/fichas', file_id)
-        )
-
-    def by_service(self, service_id):
-
-        return self._connect(
-            self._get_url('/servicios', service_id, is_file=True)
-        )
+from .chile_atiende_client import Service
+from .chile_atiende_client import File
 
 
 def create_services():
@@ -160,15 +80,9 @@ def create_files_by_services():
         file_list = []
         for file in files:
 
-            service_name = file.get('servicio')
-
-            service = ServiceModel.objects.get_or_none(
-                name=service_name
-            )
-
             data = {
                 'service': service,
-                'service_name': service_name,
+                'service_name': file.get('servicio'),
                 'title': file.get('titulo'),
                 'code': file.get('id'),
                 'date': file.get('fecha'),
